+++ settings.gradle
@@ -3,6 +3,7 @@ pluginManagement {
         google()
         mavenCentral()
         gradlePluginPortal()
+        maven { url 'https://jitpack.io'}
     }
 }
 dependencyResolutionManagement {
